// electro-pime/types/api.ts
export interface Customer {
  id: string;
  name: string;
  email?: string;
  phone: string;
  documentType: string;
  documentNumber: string;
  address?: string;
  createdAt: string;
  updatedAt: string;
}

export interface CreateCustomerDto {
  name: string;
  email?: string;
  phone: string;
  documentType: string;
  documentNumber: string;
  address?: string;
}